DCT now works. Also moved the register indices of the input channels to match the compiler's ABI (http://www.wiki.cl.cam.ac.uk/clwiki/CompArch/FAB/Strawman/ABI).

+++ src/Datatype/Word.h
@@ -41,7 +41,7 @@ public:
   Word& operator= (const Word& other);
 
   friend std::ostream& operator<< (std::ostream& os, Word const& v) {
-    os << v.data;
+    os << (int)v.data;
     return os;
   }
